#17 refactor : 리스트 스크롤 관련 기능 useAnimate 훅으로 분리
